refactored the Map module and split the layers creation into their own module
