+++ src/constants/constants.js
@@ -1,2 +1,7 @@
+import L from 'leaflet';
+
 export const WEATHER_API_URI = 'http://api.openweathermap.org/';
 export const MAX_FAV_COUNT = 3;
+export const MAP_MAX_BOUNDS = new L.LatLngBounds(new L.LatLng(-70, -180), new L.LatLng(180, 190));
+export const MAP_MAX_VISCOSITY = 0.8;
+export const MAP_MAX_ZOOM = 1;
